cms/requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.4
distlib==0.3.7
Django==4.2.4
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
environ==1.0
filelock==3.12.2
platformdirs==3.10.0
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
virtualenv==20.24.2
